Fixed the implementation to handle "*" sub-domains.
Updated the public_suffix_list.dat to the latest.
Added an `mk` script to easily rebuild the library and tools.
Allow the public_suffix_list.dat to be in the tests/... sub-folder.
Updated the tests to pass with the latest version.
Added a test to verify specific URLs (in tld_test.c).
Fixed file description at the top of two tests.
Added the .gitignore file.
Replaced many tabs with spaces so diffs as expected.
Updated the copyright year.