This is a first stab at removing Qt.
The tld_data.xml file was parsed using Qt which is heavy. Instead, I wanted
to change the format to a set of .ini files, have a compiler in the library
so one can update the files along the way without recompiling, and as a
result we can use our own parser. This is a very first stride in that
direction. It compiles, it doesn't link yet and even less compile the .ini
file at all.