Converted entities to backslash escapes.
Entities were great in XML and automatically converted by the Qt library.
Now the library parses everything so converting to something that a .ini
expects makes more sense. Using the C-like backslash notation works.

Added a note about the fact that periods (.) and backslashes (\) in domain
names are not used for hierarchy.

Started adding code to verify TLD characters. We now accept all Unicode,
only ASCII characters are checked for errors. Added means to detect errors
when adding a domain segment to a TLD definition.

The append_wc() function can generate an error if the input character is
not a valid Unicode code point. This can now easily happen since we allow
the user to enter large Unicode character codes with `"\U..."`.