Got the compiler to work properly.
This also includes the file loader, which properly reloads the data as expected.
Added a way to output the TLDs in a JSON database.
Added a string manage which has the ability to "compress" the strings (by about 3x, under 50Kb instead of over 150Kb uncompressed).
Now each TLD can be assigned any number of tags (i.e. "tag/<name>=<value>").
Changed the category & region & country name into tags.
Added a tag "language" to describe the language used for a given TLD (if not English).
Made the "nic" URL a tag.
Made the description & note tags.
Fixed the sorting of the TLDs in the compiler.
Added "tag/" in front of many variables to transform them in tags.
Marked .broadway as a "street region" now that tags are used.
Marked Japanese prefectures as such instead of "counties".
Moved the list of sources of the tld library to a variable in the CMakeLists.txt so it can be used it twice.
Added the tld_status_to_string() function.
The compiler uses the JSON to verify that what gets compiled and what gets reloaded are equivalent.
Missing: I think that the "*" support is not working in the output (I have not seen one such entry in the results...)