Fixed some invalid characters in the .ini.
I found some \u<5-digits>, when all were supposed to be \u<4-digits>. That was wrong in my old tld_data.xml file.
TLDs with large code points (over 0xFFFF) will work correctly in this newer version. Older versions would ignore the higher 4 bits. That being said, none of the TLDs use such high characters, but there were bugs in that regard in the old version.
Moved additional TLDs to specific brand .ini files.
Fixed the .www.ck exception, I had 4 'w's somehow.
Explicitly added the asterisks in all ".*.<tld>" which needed it.
I put .sci back in, but I don't think it was a "region". I found a document to have it as "science" and now that we have a .science...
Install the manual pages as is, the dh_compress is expected to take care of the compression.
Fixed the linking of the manual pages using the .links debian file.
Fixed the softlink of the "extensions" using a .links debian file.
Updated all the tests that required updates, especially when accessing the tld_data.c old structure.
Added code to grab the "category" and "country" codes and save them in the info structure.
For that purpose, I added a tld_word_to_category() function which allows me to convert a category word in a TLD_CATEGORY_... for backward compatibility.
Fixed the add_segment() by (1) verifying that if we get "*" it has a length of 1 and (2) if a character is not a basic one, then use the %XX syntax.
Fixed the input stream type to std::ifstream so that way it is not an input/output (was just std::fstream), in tld_parser.
Fixed the build of the tld_data.c, it would get built 2 or 3 times in a row each time. Now it works exactly as expected.