Started work to get libtld coverage to run.
Removed the libtld-only-CMakeLists.txt because I do not support it and the new version (2.x) should not require it anyway.
Fixed the debian/README.* files still referencing libtld-only-CMakeLists.txt
Cleaned up a few CMakeLists.txt files further.
Removed the boost library dependency. It was used for one single line of code.
Fixed the destination path to install the tlds .ini files in the compiler package.
Updated the hex tool to print out the new format (\uXXXX instead of &#XXXX;).
Added a TODO.txt file on things I can think of I'd like to appear in the project.
Fixed the path in the run_all_tests.sh so it works with the ./mk -c command (i.e. coverage).
Updated the load_tlds() of the tld_tests.c to attempt to load. Err only if both fail.
Added a free_tlds() to the tld_tests.c file so the sanitizer is happy.
Added two classes to auto-free the TLD file and strings.