Got the coverage test to pass.
Fixed one bug in the tld_test_tld_names.c where a variable was getting smashed.
Swapped a couple of tests I just added, it was meant to continue on a nullptr.
Fixed the two memcmp(), the length must match otherwise we don't even need to run the memcmp().
Make sure that we do not copy the country name if it were to overflow the available space in our tld_info structure.
Fixed a test with the tag compression code, it would eventually try the wrong way.
Made the run_all_tests.sh script work with command line options as received from the coverage script.
Added a tld_test_validate.sh script to test the validate-tld tool, this was part of the coverage before.