Trying with gdb to try to find an exception.
The tld_parser throws with an vector::_insert_size error, which can happen
if we mix two vectors together (i.e. begin() of one vector and end() of
another). It does not happen on my computer so I have to debug there.

Also, added a (currently unused) --verbose option to the tldc tool and
changed the std::cout to use std::endl to make sure that we get the message
if possible (i.e. if buffered it would not be output on an exception).