Found the culptrit. Wrong index!
However, the algorithm works on my computer (somehow) unless I suffle the
vector where I read the input files. When I do that, it failed like on the
build system (and that's how I found the invalid index). So next I'll need
to find out what's wrong in the algorithm.