Removed the optimization which breaks.
The optimization I have of the intermediate tags fails really bad. I don't
want to spend too much time on it for now. Cleaned up a few tests as an
insurance that it works as expected. The tags are properly loaded for
Macao for now. I should have a test that verifies all the tags.