Updated the .ini files with the latest public suffix list.
Fixed the compiler support for UTF-8. A variable was not being incremented.
Fixed the compiler .c file output generator that would not properly output the last line when the size is an exact multiple of 16.
Fixed the f_strings_end value which would be way larger than it was supposed to be.
Added support for a "contact" category so we can include a contact person name & email.
Changed the debhelper version as a dependency instead of the debian/compat file.
Fixed the list of dependencies for the tld_parser. The file(GLOB_RECURSE ...) is not detecting new files without a forced rebuild.
Fixed documentation about the TLD_STATUS_EXCEPTION.
Fixed the tld_test.c cat_ext() that would do random concatenations.
Made many enhancements to the tests.